+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 * refactor: add general class metadata
 * feat: support for immutable setters
 * test: immutable setter without setter on parent object error
+* test: immutable setter now works using 'wither' method
 
 ## 1.8.0